Add files via upload
I add the power law distribution with exponential decay function fit 
\begin{equation}
f(x) \propto x^{-\alpha} e^{-\beta x}
\end{equation}
but now the exponent is 0.79 not the theoretical value of 1.07
Perhaps we should change the range of data values to fit the function. Additionally, I have modified the code for the periodic boundary condition by changing if min_index > 0: to if min_index >= 0:. This ensures that the fitness of the last species is updated when the first one is selected.